+++ activerecord/lib/active_record/relation/query_methods.rb
@@ -1415,6 +1415,24 @@ module ActiveRecord
       self
     end
 
+    # Specifies whether the records should be unique per columns (only supported on PostgreSQL). For example:
+    #
+    #   User.eager_load(:bikes)
+    #   # Might return duplicate user records
+    #
+    #   User.eager_load(:bikes).distinct_on(:id)
+    #   # Returns a record per distinct users.id
+    def distinct_on(*args)
+      check_if_method_has_arguments!(:distinct_on, args)
+      spawn.distinct_on!(*args)
+    end
+
+    # Like #distinct_on, but modifies relation in place.
+    def distinct_on!(*args) # :nodoc:
+      self.distinct_on_values += args
+      self
+    end
+
     # Used to extend a scope with additional methods, either through
     # a module or through a block provided.
     #
@@ -1762,6 +1780,7 @@ module ActiveRecord
 
         arel.optimizer_hints(*optimizer_hints_values) unless optimizer_hints_values.empty?
         arel.distinct(distinct_value)
+        arel.distinct_on(*arel_columns(distinct_on_values.uniq)) unless distinct_on_values.empty?
         arel.from(build_from) unless from_clause.empty?
         arel.lock(lock_value) if lock_value